fix(#166): use intValue() for Integer comparison in computeNormalizedCut
Integer objects from Map<String, Integer>.get() compared with == only
work for values in [-128, 127] due to Java's Integer cache. For cluster
IDs > 127, the identity comparison silently fails, producing incorrect
normalized cut values.

Replace == and != with .intValue() to ensure correct primitive
comparison regardless of cluster count.

Fixes #166

+++ Gvisual/src/gvisual/GraphSpectralClusteringEngine.java
@@ -799,11 +799,11 @@ public class GraphSpectralClusteringEngine {
         for (int c = 0; c < k; c++) {
             double cut = 0, vol = 0;
             for (int i = 0; i < n; i++) {
-                boolean inC = assignment.get(vertices.get(i)) == c;
+                boolean inC = assignment.get(vertices.get(i)).intValue() == c;
                 if (inC) {
                     for (int j = 0; j < n; j++) vol += adj[i][j];
                     for (int j = 0; j < n; j++) {
-                        if (assignment.get(vertices.get(j)) != c) {
+                        if (assignment.get(vertices.get(j)).intValue() != c) {
                             cut += adj[i][j];
                         }
                     }
